requirements.txt
gevent==24.11.1
greenlet==3.1.1
msgpack==1.1.0
python-dotenv==1.0.1
setuptools==75.6.0
zope.event==5.0
zope.interface==7.2
